requirements.txt
flet==0.24.1
argparse==1.4.0
numpy==1.24.4
sounddevice==0.4.6
scipy==1.10.1
openpyxl==3.1.2
pyannote.audio==0.0.1
pytorch-lightning==1.6.0
torch==1.10.0
torchaudio==0.11.0
torchvision==0.12.0 
setuptools==58.0.4